# docs/creamoda.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "creamoda",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/host/api/v1/user/login": {
      "post": {
        "summary": "用户登陆接口",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "pwd": {
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "pwd"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "authorization": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "authorization"
                      ]
                    }
                  },
                  "required": [
                    "code",
                    "msg",
                    "data"
                  ]
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "msg": "string",
                      "data": {
                        "authorization": "Bear......"
                      }
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/host/api/v1/user/register": {
      "post": {
        "summary": "注册接口",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "pwd": {
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "pwd"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "msg",
                    "01JMDXPX95RTRSYGQP304JN3TX"
                  ]
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "msg": "string"
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/host/api/v1/user/info": {
      "get": {
        "summary": "查询用户信息接口",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "uid",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "name": {
                      "type": "string",
                      "description": "名称"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "username": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string"
                        },
                        "headPic": {
                          "type": "string"
                        },
                        "emailVerified": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "username",
                        "email",
                        "status",
                        "headPic",
                        "emailVerified"
                      ]
                    }
                  },
                  "required": [
                    "code",
                    "name",
                    "data"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/host/api/v1/user/email/verify": {
      "post": {
        "summary": "验证用户邮箱接口",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "verifyCode": {
                    "type": "string"
                  }
                },
                "required": [
                  "verifyCode"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "msg",
                    "01JMDYCP2PT9PDXQP93WF7PVA7"
                  ]
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "msg": "string"
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}